requirements.txt
fastapi 
uvicorn
python-dotenv
langchain
Cython
llama_index
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
pydantic
langchain_community
langchain_openai
langchain-text-splitters
openai
markdown
pygments
tiktoken
beautifulsoup4
pymupdf 
python-docx
faiss-cpu==1.7.2
# simsimd
pandas
openpyxl
python-pptx
httpx
azure-search-documents
azure-identity
aiohttp
clean-text
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
azure-monitor-opentelemetry
azure-monitor-events-extension 
